feat(types): make document arrays into proxies
Re: #8884
