+++ lib/types/documentarray.js
@@ -341,6 +341,70 @@ function _updateParentPopulated(arr) {
   }
 }
 
+function documentArrayProxy(arr) {
+  const proxy = new Proxy(arr, {
+    get: function(target, prop) {
+      if (prop === 'isMongooseArray' || prop === 'isMongooseArrayProxy' || prop === 'isMongooseDocumentArray') {
+        return true;
+      }
+      if (prop === '__array') {
+        return arr;
+      }
+      if (prop === 'set') {
+        return __set;
+      }
+      if (CoreDocumentArray.prototype.hasOwnProperty(prop)) {
+        return CoreDocumentArray.prototype[prop];
+      }
+      if (CoreMongooseArray.prototype.hasOwnProperty(prop)) {
+        return CoreMongooseArray.prototype[prop];
+      }
+
+      return arr[prop];
+    },
+    set: function(target, prop, value) {
+      if (typeof prop === 'string' && /^\d+$/.test(prop)) {
+        set(proxy, prop, value);
+      } else {
+        arr[prop] = value;
+      }
+
+      return true;
+    }
+  });
+
+  return proxy;
+};
+
+/*!
+ * Used as a method by array instances
+ */
+function __set(i, val, skipModified) {
+  const arr = this.__array;
+  if (skipModified) {
+    arr[i] = val;
+    return arr;
+  }
+  const value = CoreDocumentArray.prototype._cast.call(this, val, i);
+  arr[i] = value;
+  CoreDocumentArray.prototype._markModified.call(this, i);
+  return arr;
+}
+
+/**
+ * Internal `set()` logic for proxies
+ */
+function set(target, i, val, skipModified) {
+  if (skipModified) {
+    target.__array[i] = val;
+    return target;
+  }
+  const value = CoreDocumentArray.prototype._cast.call(target, val, i);
+  target.__array[i] = value;
+  CoreDocumentArray.prototype._markModified.call(target, i);
+  return target;
+}
+
 /**
  * DocumentArray constructor
  *
@@ -354,9 +418,7 @@ function _updateParentPopulated(arr) {
  */
 
 function MongooseDocumentArray(values, path, doc) {
-  // TODO: replace this with `new CoreDocumentArray().concat()` when we remove
-  // support for node 4.x and 5.x, see https://i.imgur.com/UAAHk4S.png
-  const arr = new CoreDocumentArray();
+  const arr = [];
 
   arr[arrayAtomicsSymbol] = {};
   arr[arraySchemaSymbol] = void 0;
@@ -391,7 +453,7 @@ function MongooseDocumentArray(values, path, doc) {
     }
   }
 
-  return arr;
+  return documentArrayProxy(arr);
 }
 
 /*!
